Remake SynthesisCore module in .NET Standard and begin moving files over
