--- engine/Assets/Scripts/Engine/CameraController.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: f240ee69ddf8b95499610c974e3913f6
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
