--- engine/Assets/Scripts/Engine/Interaction.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: 68a5b076e73f1f542bcd8d67c93fc9ee
-folderAsset: yes
-DefaultImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
