--- engine/Assets/Scripts/Engine/Interaction/BlankSelectable.cs
@@ -1,11 +1,0 @@
-using UnityEngine;
-
-namespace Synthesis.Simulator.Interaction
-{
-
-    public class BlankSelectable : ISelectable
-    {
-        public Vector3 Position { get; set; }
-    }
-
-}
