--- engine/Assets/Scripts/Engine/Interaction/BlankSelectable.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: a9fbf0dbb4d878447be2556edf52fe1f
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
