--- engine/Assets/Scripts/Engine/Interaction/ISelectable.cs
@@ -1,11 +1,0 @@
-﻿using System.Collections;
-using System.Collections.Generic;
-using UnityEngine;
-
-namespace Synthesis.Simulator.Interaction
-{
-    public interface ISelectable
-    {
-        Vector3 Position { get; }
-    }
-}
