--- engine/Assets/Scripts/Engine/Interaction/ISelectable.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: fce6a70d08e6bad41a263fb90d99f7f0
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
