--- engine/Assets/Scripts/Engine/Simulator.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: da819d533030c774c957d4e494df6978
-folderAsset: yes
-DefaultImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
