--- engine/Assets/Scripts/Engine/Simulator/Robot.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 0a1a64d7a486e7946aaa149594086166
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
