+++ modules/SynthesisCore/SampleSystem.cs
