+++ modules/SynthesisCore/SynthesisCore.cs
