+++ modules/modules.sln
@@ -5,7 +5,7 @@ VisualStudioVersion = 16.0.30204.135
 MinimumVisualStudioVersion = 10.0.40219.1
 Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Controller", "Controller\Controller.csproj", "{4B6DDC63-08B5-4743-B3CA-41C20C1741AB}"
 EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "SynthesisCore", "SynthesisCore\SynthesisCore.csproj", "{FD33E89C-FE8D-490B-82B8-7063B707643E}"
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "SynthesisCore", "SynthesisCore\SynthesisCore.csproj", "{B4A3AF90-D904-454E-BAEE-6B3497A36402}"
 EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
@@ -17,10 +17,10 @@ Global
 		{4B6DDC63-08B5-4743-B3CA-41C20C1741AB}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{4B6DDC63-08B5-4743-B3CA-41C20C1741AB}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{4B6DDC63-08B5-4743-B3CA-41C20C1741AB}.Release|Any CPU.Build.0 = Release|Any CPU
-		{FD33E89C-FE8D-490B-82B8-7063B707643E}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{FD33E89C-FE8D-490B-82B8-7063B707643E}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{FD33E89C-FE8D-490B-82B8-7063B707643E}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{FD33E89C-FE8D-490B-82B8-7063B707643E}.Release|Any CPU.Build.0 = Release|Any CPU
+		{B4A3AF90-D904-454E-BAEE-6B3497A36402}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{B4A3AF90-D904-454E-BAEE-6B3497A36402}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{B4A3AF90-D904-454E-BAEE-6B3497A36402}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{B4A3AF90-D904-454E-BAEE-6B3497A36402}.Release|Any CPU.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
